# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LaunchMailer API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /upload/request-url:
    post:
      operationId: requestUploadUrl
      tags: [upload]
      summary: Request a presigned URL for image upload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlInput"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"

  # ─── AI GENERATE ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UploadUrlInput:
      type: object
      required: [filename, contentType]
      properties:
        filename:
          type: string
        contentType:
          type: string
          enum: [image/jpeg, image/jpg, image/png, image/webp]
    UploadUrlResponse:
      type: object
      required: [uploadUrl, publicUrl, objectKey]
      properties:
        uploadUrl:
          type: string
          description: Presigned PUT URL for uploading
        publicUrl:
          type: string
          description: Public URL after upload
        objectKey:
          type: string

    # ─── AI GENERATE ──────────────────────────────────────────────────────────
...